requirements.txt
Django==3.0.6
psycopg2==2.8.5
djangorestframework==3.11.0
django-environ==0.4.5
gunicorn==20.0.4